Move actions and selectors to private.
