+++ packages/core-data/README.md
@@ -248,19 +248,6 @@ _Returns_
 
 -   `Object`: Action object.
 
-### receiveRegisteredPostMeta
-
-Returns an action object used in signalling that the registered post meta fields for a post type have been received.
-
-_Parameters_
-
--   _postType_ `string`: Post type slug.
--   _registeredPostMeta_ `Object`: Registered post meta.
-
-_Returns_
-
--   `Object`: Action object.
-
 ### receiveRevisions
 
 Action triggered to receive revision items.
@@ -755,19 +742,6 @@ _Parameters_
 _Returns_
 
 -   A value whose reference will change only when an edit occurs.
-
-### getRegisteredPostMeta
-
-Returns the registered post meta fields for a given post type.
-
-_Parameters_
-
--   _state_ `State`: Data state.
--   _postType_ `string`: Post type.
-
-_Returns_
-
--   Registered post meta fields.
 
 ### getRevision
