+++ packages/core-data/src/selectors.ts
@@ -1527,15 +1527,3 @@ export const getRevision = createSelector(
 		];
 	}
 );
-
-/**
- * Returns the registered post meta fields for a given post type.
- *
- * @param state    Data state.
- * @param postType Post type.
- *
- * @return Registered post meta fields.
- */
-export function getRegisteredPostMeta( state: State, postType: string ) {
-	return state.registeredPostMeta?.[ postType ] ?? {};
-}
